requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.16.0
onnx>=1.14.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
pycocotools>=2.0
tqdm>=4.65.0
pyyaml>=6.0
Pillow>=10.0.0
albumentations>=1.3.0
torchmetrics>=1.0.0
